Sync MTLDevice/NSArray patterns, use new() syntax, fix nativePtr naming, fix uint/ulong type mapping

+++ Metal.NET/Metal/MTL4CommandQueue.cs
@@ -37,9 +37,9 @@ public class MTL4CommandQueue(nint nativePtr) : NativeObject(nativePtr)
         ObjectiveCRuntime.MsgSend(NativePtr, MTL4CommandQueueBindings.SignalDrawable, drawable.NativePtr);
     }
 
-    public void SignalEvent(MTLEvent @event, nuint value)
+    public void SignalEvent(MTLEvent @event, ulong value)
     {
-        ObjectiveCRuntime.MsgSend(NativePtr, MTL4CommandQueueBindings.SignalEvent, @event.NativePtr, value);
+        ObjectiveCRuntime.MsgSend(NativePtr, MTL4CommandQueueBindings.SignalEvent, @event.NativePtr, (nuint)value);
     }
 
     public void UpdateBufferMappings(MTLBuffer buffer, MTLHeap heap, MTL4UpdateSparseBufferMappingOperation operations, nuint count)
@@ -52,9 +52,9 @@ public class MTL4CommandQueue(nint nativePtr) : NativeObject(nativePtr)
         ObjectiveCRuntime.MsgSend(NativePtr, MTL4CommandQueueBindings.UpdateTextureMappings, texture.NativePtr, heap.NativePtr, operations, count);
     }
 
-    public void Wait(MTLEvent @event, nuint value)
+    public void Wait(MTLEvent @event, ulong value)
     {
-        ObjectiveCRuntime.MsgSend(NativePtr, MTL4CommandQueueBindings.Wait, @event.NativePtr, value);
+        ObjectiveCRuntime.MsgSend(NativePtr, MTL4CommandQueueBindings.Wait, @event.NativePtr, (nuint)value);
     }
 
     public void Wait(MTLDrawable drawable)
